pyproject.toml
[project]
name = "address-geocoder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "adbc-driver-manager>=1.8.0",
    "adbc-driver-postgresql>=1.8.0",
    "black>=25.9.0",
    "ipython>=8.18.1",
    "pandas>=2.2",
    "passyunk",
    "petl>=1.7.17",
    "polars>=1.34.0",
    "psycopg2-binary>=2.9",
    "pyarrow>=14",
    "pytest>=8.4.2",
    "pyyaml>=6.0.3",
    "retrying>=1.4.2",
    "sqlalchemy>=2.0",
]

[tool.uv.sources]
passyunk = { git = "https://github.com/CityOfPhiladelphia/passyunk" }
